[TM-1314] edit routes and change table insights
