+++ src/components/elements/Table/TableVariants.ts
@@ -145,7 +145,7 @@ export const VARIANT_TABLE_DASHBOARD_COUNTRIES_MODAL = {
 export const VARIANT_TABLE_DASHBOARD = {
   table: "border-collapse",
   name: "border-airtable",
-  tableWrapper: "border border-neutral-200 rounded-lg overflow-auto w-full max-h-[calc(100vh-252px)]",
+  tableWrapper: "border border-neutral-200 rounded-lg overflow-auto w-full max-h-[calc(100vh-273px)]",
   trHeader: "bg-neutral-150 sticky top-0 z-auto",
   thHeader:
     "first:pl-4 first:pr-2 last:pl-2 last:pr-4 border-y border-neutral-200 text-14-semibold whitespace-normal px-1.5 border-t-0 py-4",
@@ -158,7 +158,7 @@ export const VARIANT_TABLE_DASHBOARD = {
 export const VARIANT_TABLE_AIRTABLE_DASHBOARD = {
   table: "border-collapse",
   name: "border-airtable",
-  tableWrapper: "border border-neutral-200 rounded-lg overflow-auto w-full max-h-[calc(100vh-283px)]",
+  tableWrapper: "border border-neutral-200 rounded-lg overflow-auto w-full max-h-[calc(100vh-219px)]",
   trHeader: "bg-neutral-150 sticky top-0 z-auto",
   thHeader:
     "first:pl-4 first:pr-2 last:pl-2 last:pr-4 border-y border-neutral-200 text-14-semibold whitespace-normal px-1.5 border-t-0 py-4",
