+++ src/components/generic/Sidebar/Sidebar.tsx
@@ -11,15 +11,18 @@ const Sidebar = () => {
 
   return (
     <div className="flex flex-col items-center gap-8 bg-blueCustom-700 p-3 text-darkCustom-200">
-      <a className="mb-10 mt-4 cursor-pointer" href="/home">
+      <a className="mt-4 mb-10 cursor-pointer" href="/home">
         <Icon name={IconNames.TERRAFUND_lOGO_MINI} />
       </a>
       <Tooltip content="DASHBOARDS" placement="right">
         <a
           className={classNames("flex cursor-pointer flex-col items-center gap-1", {
-            "text-white": router.asPath === "/dashboard/project" || router.asPath.includes("/dashboard/country")
+            "text-white":
+              router.asPath.includes("/dashboard") &&
+              !router.asPath.includes("/dashboard/project-list") &&
+              !router.asPath.includes("/dashboard/project-insights")
           })}
-          href="/dashboard/country"
+          href="/dashboard"
         >
           <Icon name={IconNames.DASHBOARDS} className="h-8 w-8" />
           <Text variant={"text-8"}>DASHBOARDS</Text>
